requirements.txt
# Core
psycopg2-binary==2.9.9
redis==5.0.1
pandas==2.2.1
numpy==1.26.4

# Data loading
requests==2.31.0
yfinance==0.2.36

# Backtest
scipy==1.12.0
python-dateutil==2.9.0

# RAG Pipeline (Phase 3)
pdfplumber==0.11.0
chromadb==0.4.22
sentence-transformers==2.3.1
anthropic==0.18.1
pytesseract==0.3.10
Pillow==10.2.0

# Monitoring (Phase 5)
# python-telegram-bot==21.0

# Execution (Phase 4)
# kiteconnect==5.0.1

# FII Pipeline (Phase 5)
# playwright==1.41.2
